requirements.txt
pytrends>=4.9.2
anthropic>=0.49.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=22.0.0
